requirements.txt
pdfplumber
pandas
pyyaml
beautifulsoup4